+++ src/Alpaca/views/__init__.py
@@ -1,0 +1,9 @@
+# Import here all your views
+from others import *
+
+from authentication import *
+from user_management import *
+
+from activity import *
+from session import *
+
